server/requirements.txt
# TianShu - Dependencies

# Core
Flask==3.0.0
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36

# API & Serialization
Flask-RESTX==1.3.0
marshmallow==3.20.1

# Authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.1

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Validation
email-validator==2.1.0

# CORS
Flask-Cors==4.0.0

# WebSocket
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.1
eventlet==0.40.4

# Database Migration
Flask-Migrate==4.0.5

# Utils
python-dateutil==2.8.2

# Scheduler
APScheduler==3.10.4

# Production Server
gunicorn==21.2.0
